move config_type.py
