+++ tests/test_fin_class.py
@@ -31,7 +31,7 @@ import os
 import tempfile
 
 from dummy_machine import DummyMachine
-from tuna.config_type import ConfigType
+from tuna.miopen.utils.config_type import ConfigType
 from multiprocessing import Value, Lock, Queue
 from tuna.miopen.utils.metadata import LOG_TIMEOUT
 from tuna.miopen.worker.fin_class import FinClass
