+++ tests/test_update_golden.py
@@ -28,7 +28,7 @@ import sys
 from tuna.miopen.modules.update_golden import merge_golden_entries, get_fdb_entries, create_perf_table, verify_no_duplicates, latest_golden_v
 from tuna.miopen.db.tables import MIOpenDBTables
 from tuna.dbBase.sql_alchemy import DbSession
-from tuna.config_type import ConfigType
+from tuna.miopen.utils.config_type import ConfigType
 from tuna.miopen.db.miopen_tables import ConvolutionGolden
 from tuna.miopen.db.find_db import ConvolutionFindDB
 from utils import add_test_session, DummyArgs
