+++ tuna/miopen/driver/convolution.py
@@ -36,7 +36,7 @@ from tuna.miopen.utils.metadata import CONV_2D_DEFAULTS, SUPPORTED_CONV_CMDS, PR
 from tuna.miopen.utils.metadata import CONV_3D_DEFAULTS, TENSOR_COLS, TABLE_COLS_CONV_MAP, TENSOR_PRECISION
 from tuna.miopen.utils.metadata import DIRECTION, DIR_MAP, CONV_SKIP_ARGS, INVERS_DIR_MAP
 from tuna.miopen.utils.parsing import get_fd_name, conv_arg_valid, get_fds_from_cmd
-from tuna.config_type import ConfigType
+from tuna.miopen.utils.config_type import ConfigType
 
 LOGGER = setup_logger('driver_conv')
